+++ lib/zip_dirent.c
@@ -37,6 +37,7 @@
 #include <string.h>
 #include <sys/types.h>
 #include <time.h>
+#include <zlib.h>
 
 #include "zipint.h"
 
@@ -127,6 +128,7 @@ _zip_cdir_write(zip_t *za, const zip_filelist_t *filelist, zip_uint64_t survivor
     zip_uint64_t i;
     bool is_zip64;
     int ret;
+    zip_uint32_t cdir_crc;
 
     if ((off = zip_source_tell_write(za->src)) < 0) {
         zip_error_set_from_source(&za->error, za->src);
@@ -135,6 +137,11 @@ _zip_cdir_write(zip_t *za, const zip_filelist_t *filelist, zip_uint64_t survivor
     offset = (zip_uint64_t)off;
 
     is_zip64 = false;
+
+    if (ZIP_WANT_TORRENTZIP(za)) {
+        cdir_crc = (zip_uint32_t)crc32(0, NULL, 0);
+        za->write_crc = &cdir_crc;
+    }
 
     for (i = 0; i < survivors; i++) {
         zip_entry_t *entry = za->entry + filelist[i].idx;
@@ -145,15 +152,17 @@ _zip_cdir_write(zip_t *za, const zip_filelist_t *filelist, zip_uint64_t survivor
             is_zip64 = true;
     }
 
+    za->write_crc = NULL;
+
     if ((off = zip_source_tell_write(za->src)) < 0) {
         zip_error_set_from_source(&za->error, za->src);
         return -1;
     }
     size = (zip_uint64_t)off - offset;
 
-    if (offset > ZIP_UINT32_MAX || survivors > ZIP_UINT16_MAX)
+    if (offset > ZIP_UINT32_MAX || survivors > ZIP_UINT16_MAX) {
         is_zip64 = true;
-
+    }
 
     if ((buffer = _zip_buffer_new(buf, sizeof(buf))) == NULL) {
         zip_error_set(&za->error, ZIP_ER_MEMORY, 0);
@@ -186,7 +195,13 @@ _zip_cdir_write(zip_t *za, const zip_filelist_t *filelist, zip_uint64_t survivor
 
     comment = za->comment_changed ? za->comment_changes : za->comment_orig;
 
-    _zip_buffer_put_16(buffer, (zip_uint16_t)(comment ? comment->length : 0));
+    if (ZIP_WANT_TORRENTZIP(za)) {
+        _zip_buffer_put_16(buffer, TORRENTZIP_SIGNATURE_LENGTH + TORRENTZIP_CRC_LENGTH);
+    }
+    else {
+        _zip_buffer_put_16(buffer, (zip_uint16_t)(comment ? comment->length : 0));
+    }
+
 
     if (!_zip_buffer_ok(buffer)) {
         zip_error_set(&za->error, ZIP_ER_INTERNAL, 0);
@@ -201,7 +216,15 @@ _zip_cdir_write(zip_t *za, const zip_filelist_t *filelist, zip_uint64_t survivor
 
     _zip_buffer_free(buffer);
 
-    if (comment) {
+    if (ZIP_WANT_TORRENTZIP(za)) {
+        char torrentzip_comment[TORRENTZIP_SIGNATURE_LENGTH + TORRENTZIP_CRC_LENGTH + 1];
+        snprintf(torrentzip_comment, sizeof(torrentzip_comment), TORRENTZIP_SIGNATURE "%08X", cdir_crc);
+
+        if (_zip_write(za, torrentzip_comment, strlen(torrentzip_comment)) < 0) {
+            return -1;
+        }
+    }
+    else if (comment != NULL) {
         if (_zip_write(za, comment->raw, comment->length) < 0) {
             return -1;
         }
@@ -902,7 +925,13 @@ _zip_dirent_write(zip_t *za, zip_dirent_t *de, zip_flags_t flags) {
         _zip_buffer_put_16(buffer, (zip_uint16_t)de->comp_method);
     }
 
-    _zip_u2d_time(de->last_mod, &dostime, &dosdate);
+    if (ZIP_WANT_TORRENTZIP(za)) {
+        dostime = 0xbc00;
+        dosdate = 0x2198;
+    }
+    else {
+        _zip_u2d_time(de->last_mod, &dostime, &dosdate);
+    }
     _zip_buffer_put_16(buffer, dostime);
     _zip_buffer_put_16(buffer, dosdate);
 
@@ -938,12 +967,15 @@ _zip_dirent_write(zip_t *za, zip_dirent_t *de, zip_flags_t flags) {
     }
 
     _zip_buffer_put_16(buffer, _zip_string_length(de->filename));
-    /* TODO: check for overflow */
-    ef_total_size = (zip_uint32_t)_zip_ef_size(de->extra_fields, flags) + (zip_uint32_t)_zip_ef_size(ef, ZIP_EF_BOTH);
+    ef_total_size = (zip_uint32_t)_zip_ef_size(ef, ZIP_EF_BOTH);
+    if (!ZIP_WANT_TORRENTZIP(za)) {
+        /* TODO: check for overflow */
+        ef_total_size += (zip_uint32_t)_zip_ef_size(de->extra_fields, flags);
+    }
     _zip_buffer_put_16(buffer, (zip_uint16_t)ef_total_size);
 
     if ((flags & ZIP_FL_LOCAL) == 0) {
-        _zip_buffer_put_16(buffer, _zip_string_length(de->comment));
+        _zip_buffer_put_16(buffer, ZIP_WANT_TORRENTZIP(za) ? 0 : _zip_string_length(de->comment));
         _zip_buffer_put_16(buffer, (zip_uint16_t)de->disk_number);
         _zip_buffer_put_16(buffer, de->int_attrib);
         _zip_buffer_put_32(buffer, de->ext_attrib);
@@ -982,13 +1014,13 @@ _zip_dirent_write(zip_t *za, zip_dirent_t *de, zip_flags_t flags) {
         }
     }
     _zip_ef_free(ef);
-    if (de->extra_fields) {
+    if (de->extra_fields && !ZIP_WANT_TORRENTZIP(za)) {
         if (_zip_ef_write(za, de->extra_fields, flags) < 0) {
             return -1;
         }
     }
 
-    if ((flags & ZIP_FL_LOCAL) == 0) {
+    if ((flags & ZIP_FL_LOCAL) == 0 && !ZIP_WANT_TORRENTZIP(za)) {
         if (de->comment) {
             if (_zip_string_write(za, de->comment) < 0) {
                 return -1;
@@ -1153,4 +1185,23 @@ _zip_dirent_apply_attributes(zip_dirent_t *de, zip_file_attributes_t *attributes
     if ((changed & ZIP_DIRENT_ATTRIBUTES) == 0 && (attributes->valid & ZIP_FILE_ATTRIBUTES_HOST_SYSTEM)) {
         de->version_madeby = (de->version_madeby & 0xff) | (zip_uint16_t)(attributes->host_system << 8);
     }
+}
+
+
+/* _zip_dirent_torrent_normalize(de);
+   Set values suitable for torrentzip.
+*/
+
+void zip_dirent_torrentzip_normalize(zip_dirent_t *de) {
+    de->version_madeby = 0;
+    de->version_needed = 20; /* 2.0 */
+    de->bitflags = 2; /* maximum compression */
+    de->comp_method = ZIP_CM_DEFLATE;
+    de->compression_level = TORRENTZIP_COMPRESSION_FLAGS;
+    de->disk_number = 0;
+    de->int_attrib = 0;
+    de->ext_attrib = 0;
+
+    /* last_mod, extra_fields, and comment are normalized in zip_dirent_write() directly */
+
 }