+++ NEWS.md
@@ -8,6 +8,7 @@
 * CMake: add detection of mbedTLS version 3.
 * Use ISO C secure library functions, if available.
 * Fix reading/writing compressed data with buffers > 4GiB.
+* Restore support for torrentzip.
 
 
 # 1.9.2 [2022-06-28]